vnc server: Fix that vnc_updater thread exited caused by readed a null data
VELAPLATFO-7700

When updating the full-screen data, sq_ init() will clear the updqueue and add just one new full-screen data to the updqueue. So when the vnc_updater thread is awakened, it may read a null data due to multiple reading, leading to vnc_updater thread exited , which is not expected.

Signed-off-by: jianglianfang <jianglianfang@xiaomi.com>
Change-Id: I8be001ccdc626665d31f48c110cba62ef083601d